[refactor] DataSource: move existing Ktor endpoint code to /restapi
