+++ composeApp/src/commonTest/kotlin/com/rwmobi/kunigami/data/source/network/AccountEndpointTest.kt
@@ -7,6 +7,7 @@
 
 package com.rwmobi.kunigami.data.source.network
 
+import com.rwmobi.kunigami.data.source.network.restapi.AccountEndpoint
 import com.rwmobi.kunigami.domain.exceptions.HttpException
 import com.rwmobi.kunigami.test.samples.GetAccountSampleData
 import io.ktor.client.HttpClient
