+++ composeApp/src/commonTest/kotlin/com/rwmobi/kunigami/data/source/network/ElectricityMeterPointsEndpointTest.kt
@@ -7,6 +7,7 @@
 
 package com.rwmobi.kunigami.data.source.network
 
+import com.rwmobi.kunigami.data.source.network.restapi.ElectricityMeterPointsEndpoint
 import com.rwmobi.kunigami.domain.exceptions.HttpException
 import com.rwmobi.kunigami.test.samples.GetConsumptionSampleData
 import io.ktor.client.HttpClient
