+++ composeApp/src/commonTest/kotlin/com/rwmobi/kunigami/data/source/network/ProductsEndpointTest.kt
@@ -7,6 +7,7 @@
 
 package com.rwmobi.kunigami.data.source.network
 
+import com.rwmobi.kunigami.data.source.network.restapi.ProductsEndpoint
 import com.rwmobi.kunigami.domain.exceptions.HttpException
 import com.rwmobi.kunigami.test.samples.GetDayUnitRatesSampleData
 import com.rwmobi.kunigami.test.samples.GetNightUnitRatesSampleData
